+++ src/pseudotty.hpp
@@ -37,8 +37,8 @@ public:
             unsigned short ws_xpixel;   /* unused */
             unsigned short ws_ypixel;   /* unused */
         } winsize = {};
-        winsize.ws_row = rows;
-        winsize.ws_col = cols;
+        winsize.ws_row = static_cast<unsigned short>(rows);
+        winsize.ws_col = static_cast<unsigned short>(cols);
         ioctl(master, TIOCSWINSZ, &winsize);
     }
